requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.27.0
scikit-learn==1.6.1
Pillow==10.0.0
joblib==1.3.2
